integrate ApplyBookStudy into ApplyBook
