--- Assets/Scripts/Study/ApplyBookStudy.cs.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: 41c7e4c8a883a46e7a86856cb36e3610
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - BookPrefab: {fileID: 7874733584314137426, guid: 267acd631a4a6424cae919ab41b47de9, type: 3}
-  - bookShelf: {instanceID: 0}
-  - Data: {instanceID: 0}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
